requirements.txt
asttokens==2.2.1
blis==0.7.9
catalogue==2.0.8
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
graphviz==0.20.1
icecream==2.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.7.0
murmurhash==1.0.9
mutagen==1.46.0
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
pydantic==1.10.5
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pypiwin32==223
pytextrank==3.2.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-pexels==1.0
pyttsx3==2.90
pytz==2022.7.1
pywin32==305
requests==2.28.1
scipy==1.10.1
shotstack-sdk==0.2.5
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.13
wasabi==1.1.1
Werkzeug==2.2.2
wincertstore==0.2
